Allocate an exception stack for all modes.
We don't need to handle any exceptions yet, but allocate a stack that's shared
for all exception modes so that it can be used to implement a crash handler.

Also, we are now leaving the CPU in system mode after enabling the MMU, instead
of supervisor mode. This makes exception handling slightly easier and means it
will be possible to make SWI calls in the "kernel" later.

Also tidy up some comments.