Implement vectors that go to a crash handler.
Vectors are mapped and implemented, and when invoked just jump to a crash
handler which will dump the user/system registers at the point of the crash,
as well as CPSR in the exception mode so you can see which exception it was.
This isn't really enough information (no MMU status, etc) but it's a start.