Rename/refactor serial implementation.
The "generic" memory mapped UART is not as generic as I'd hoped: rename
it to uart8250 to reflect what it actually is, an interface for 8250ish
UARTs. Also, split the common boot debug code (handling of strings/ints)
into a separate file so it can be reused between serial drivers.