Fix line endings and refactor serial code.
Instead of handling entire arrays of char in the serial drivers, just
handle one byte at a time, since this is a debug port and we don't care
about efficiency. This enables all the line-ending modification code to
be in just one place. We now output CRLF (\r\n) for line endings, and
read either CR or CRLF (\r or \r\n). Terminals sending just LF won't
work, but that should be rare and we can't accept all three without
storing state.